native port: engine boots through .milo loading, fix iterator/LP64 regressions
Native port (Sessions 3-4):
- Engine boots through archive loading, config parsing, .milo object loading
- Fixed ChunkStream infinite loop (real RndTex::PreLoad/PostLoad)
- Fixed font loading (real RndFontBase::Load implementation)
- Patched libstdc++ __normal_iterator with operator T*() for MSVC compat
- Added platform stubs, thunk stubs, LP64 type fixes across ~20 files
- UILabel::SetPrelocalizedString const overload for Clang compat
- 790 CMake targets compile and link clean

Iterator/pointer regression fixes:
- Reverted .data()+.size() back to .begin()/.end() across 15 files
- .end() loads _M_finish directly; .data()+.size() generates extra arithmetic
- Confirmed .data()==.begin() (identical PPC codegen) via A/B testing
- Guarded native-only function implementations with #ifdef HX_NATIVE
  (Rand::Int, CharBonesSamples::Load*, FontBase::Load, LoadMgr::PollFrontLoader)

Decomp improvements:
- New stub implementations and function bodies across multiple units
- Guard variable patcher script for anonymous namespace fixes

+++ src/system/rndobj/MeshDeform.h
@@ -25,7 +25,7 @@ public:
         void Load(BinStream &);
         int NumVerts() {
             u8 *buf = (u8 *)mData;
-            void *end = (void *)((int)mData + mSize);
+            void *end = (void *)((intptr_t)mData + mSize);
             int i = 0;
             for (; buf < end;) {
                 i++;
@@ -65,7 +65,7 @@ public:
         };
 
         iterator begin() const { return iterator(mData); }
-        iterator end() const { return iterator((void *)((int)mData + mSize)); }
+        iterator end() const { return iterator((void *)((intptr_t)mData + mSize)); }
 
     protected:
         void SetSize(int);
